refactor(lib): Extract common functionality into lib crate file_server_core
