+++ Cargo.toml
@@ -3,6 +3,14 @@ name = "file-server"
 version = "0.1.0"
 edition = "2021"
 
+[[bin]]
+name = "file_server"
+path = "src/main.rs"
+
+[lib]
+name = "file_server_core"
+path = "src/lib/lib.rs"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
