+++ src/file_manager/handlers.rs
@@ -1,23 +1,26 @@
 use std::fs;
 
+use actix_web::{
+    get,
+    http::header::ContentType,
+    web::{Data, Path, Query},
+    HttpResponse, Responder,
+};
 use askama::Template;
-use actix_web::{get, http::header::ContentType, HttpResponse, Responder, web::{Data, Path, Query}};
-use log::{info, debug};
+use file_server_core::*;
+use log::{debug, info};
 
-use crate::{file_manager::templates::{HomePageTemplate, ProgramListTemplate}, configs::ServerConfigs, file_server::{handlers::{DirectoryStructureQuery, get_directory_structure_recursive, get_directory_structure}, models::Directory}};
+use crate::configs::ServerConfigs;
+use crate::file_manager::templates::{HomePageTemplate, ProgramListTemplate};
+use crate::file_server::handlers::DirectoryStructureQuery;
 
 const CSS_FILE: &'static [u8] = include_bytes!("../../public/css/main.css");
 
 #[get("/")]
 pub async fn home_page() -> impl Responder {
     info!("Getting home page");
-    let css_content = String::from_utf8(Vec::from(CSS_FILE))
-        .unwrap_or("".to_string());
-    let template = HomePageTemplate {
-        css_content,
-    }
-    .render()
-    .unwrap();
+    let css_content = String::from_utf8(Vec::from(CSS_FILE)).unwrap_or("".to_string());
+    let template = HomePageTemplate { css_content }.render().unwrap();
 
     HttpResponse::Ok()
         .insert_header(ContentType::html())
@@ -36,7 +39,7 @@ pub async fn favicon() -> impl Responder {
 pub async fn directory_structure_template(
     configs: Data<ServerConfigs>,
     path: Path<String>,
-    query: Query<DirectoryStructureQuery>
+    query: Query<DirectoryStructureQuery>,
 ) -> impl Responder {
     info!("Getting directory structure for path: {}", path);
     debug!("Query: {:?}", query);
@@ -47,14 +50,13 @@ pub async fn directory_structure_template(
     let metadata = match fs::metadata(&root_dir_path) {
         Ok(metadata) => metadata,
         Err(_) => {
-        return HttpResponse::BadRequest()
-            .body(format!("Failed to get metadata for: {:?}", root_dir_path))
+            return HttpResponse::BadRequest()
+                .body(format!("Failed to get metadata for: {:?}", root_dir_path))
         }
     };
 
     if !metadata.is_dir() {
-        return HttpResponse::BadRequest()
-            .body(format!("{:?} is not a directory", root_dir_path));
+        return HttpResponse::BadRequest().body(format!("{:?} is not a directory", root_dir_path));
     };
 
     let name = match root_dir_path.file_name() {
@@ -69,24 +71,18 @@ pub async fn directory_structure_template(
     };
 
     let get_dir_structure_result = match query.recursive {
-        Some(recursive) if recursive => {
-            get_directory_structure_recursive(&mut base_dir)
-        },
+        Some(recursive) if recursive => get_directory_structure_recursive(&mut base_dir),
         _ => get_directory_structure(&mut base_dir),
     };
-    
+
     match get_dir_structure_result {
         Ok(_) => {
             base_dir.sanitize_path(&configs.base_dir.to_string_lossy());
-            let template = ProgramListTemplate {
-                base_dir
-            }
-            .render()
-            .unwrap();
+            let template = ProgramListTemplate { base_dir }.render().unwrap();
             return HttpResponse::Ok()
                 .insert_header(ContentType::html())
                 .body(template);
-        },
+        }
         Err(err) => return HttpResponse::BadRequest().body(err.to_string()),
     }
 }
